# crm/swagger.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:5000/api/v1/",
      "description": "Devlopment server (uses test data)",
      "variables": {}
    },
    {
      "url": "http://13.126.233.224/api/v1/",
      "description": "Production server (uses live data)",
      "variables": {},
      "basePath": ""
    }
  ],
  "info": {
    "version": "v1",
    "title": "CRM API",
    "description": "This api was developed  by eng : Ahmed Ali  ✌️✌️ ,for protoqit company, @copyright protoqit 2021",
    "termsOfService": "",
    "contact": {},
    "license": { "name": "" }
  },
...
  "paths": {
...
          "/users/account/recover": {
            "post": {
              "summary": "Accunt  recover ",
              "operationId": "recover",
              "description": "you can use this end point to  recover accunt by give the email and send tokent to email",
              "parameters": [],
              "responses": { "200": { "description": "", "headers": {} } },
              "requestBody": {
                "required": true,
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "example": { "email": "ahmedshekh@student.aast.edu" }
                    },
                    "example": "{\r\n    \"email\":\"ahmedshekh@student.aast.edu\"\r\n}"
                  }
                }
              },
              "tags": ["User"]
            }
          },
          "/users/account/token/check": {
            "post": {
              "summary": "Check token is valid or not valid",
              "operationId": "checkToken",
              "description": "you can use this end point to  check if the token is valid or not valid",
              "parameters": [],
              "responses": { "200": { "description": "", "headers": {} } },
              "requestBody": {
                "required": true,
                "content": {
                  "application/json": {
                    "schema": { "type": "object", "example": { "token": "d4a1ec" } },
                    "example": "{\r\n    \"token\":\"d4a1ec\"\r\n}"
                  }
                }
              },
              "tags": ["User"]
            }
          },
          "/users/account/password/reset": { 
            "post": {
              "summary": "Reset password",
              "operationId": "resetPassword",
              "description": "you can use this end point to  reset password , you should have a tokent to confirm this operation ",
              "parameters": [],
              "responses": { "200": { "description": "", "headers": {} } },
              "requestBody": {
                "required": true,
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "example": { "token": "f9a81d", "password": "#Ahmed97" }
                    },
                    "example": "{\r\n    \"token\":\"f9a81d\",\r\n    \"password\": \"#Ahmed97\"\r\n}"
                  }
                }
              },
              "tags": ["User"]
            }
          },
...
          "/super/user/profile/{userId}": {
            "patch": {
              "summary": "update UserProfile",
              "operationId": "updateUserProfile",
              "parameters": [
                {
                  "name": "userId",
                  "in": "path",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "string" }
                }
              ],
              "responses": { "200": { "description": "", "headers": {} } },
              "requestBody": {
                "required": true,
                "content": {
                  "application/json": {
                    "schema": {
                      "type": "object",
                      "example": {
                        "phoneNumber": "01000414736",
                        "lastName": "aaaa",
                        "role": "addmin"
                      }
                    },
                    "example": "{\r\n     \"phoneNumber\": \"01000414736\",\r\n    \"lastName\":\"aaaa\",\r\n    \"role\":\"addmin\"\r\n   \r\n   \r\n \r\n    \r\n}"
                  }
                }
              },
              "tags": ["Super Admin"]
            },
            "get": {
              "summary": "get  User Profile Data",
              "operationId": "get user profile data",
              "parameters": [
                {
                  "name": "userId",
                  "in": "path",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "string" }
                }
              ],
              "responses": { "200": { "description": "", "headers": {} } },
            
              "tags": ["Super Admin"]
            }
          },
...
          "/client/file/{folderId}": {
            "get": {
              "summary": "getAllFile",
              "operationId": "getAllFile",
              "parameters": [
                {
                  "name": "pageNumber",
                  "in": "query",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "number" }
                },
                {
                  "name": "pageSize",
                  "in": "query",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "number" }
                },
                {
                  "name": "folderId",
                  "in": "path",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "string" }
                }
              ],
              "responses": { "200": { "description": "", "headers": {} } },
              "tags": ["File"]
            }
            ,"post": {
              "summary": "upload File",
              "operationId": "uploadFile",
              "parameters": [
                {
                  "name": "folderId",
                  "in": "path",
                  "required": true,
                  "style": "form",
                  "schema": { "type": "string" }
                }
              ],
              "responses": { "200": { "description": "", "headers": {} } },
              "requestBody": {
                "required": true,
                "content": {
                  "multipart/form-data": {
                    "schema": {
                      "type": "object",
                      "required": ["file"],
                      "properties": {
                        "file": { "type": "string", "format": "base64" }
                      }
                    }
                   
                  }
                }
              },
              "tags": ["File"]
            }

          },
...
  },
  "components": {
...
        "securitySchemes": {
          "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-auth-token",
            "scheme": "bearer",
            "description": "Enter JWT token",
            "in": "header"
          }
        }
      },
      "security": [{ "ApiKeyAuth": [] }],
...
}